Moved use_runner_connection to the hook module instead of it's own
